+++ .github/docker-release.sh
@@ -2,7 +2,7 @@
 
 set -e
 
-SERVICES=(issue credit basketserver)
+SERVICES=(issue credit basket)
 VERSION=$(echo "$SOURCE_TAG" | cut -c 2-) # Remove v from version
 echo "Building and deploying docker images with version: $VERSION"
 
@@ -11,7 +11,7 @@ for SERVICE in "${SERVICES[@]}"; do
 
   echo "Building docker images for ${SERVICE}-service."
 
-  ./gradlew ":${SERVICE}:bootBuildImage"
+  ./gradlew ":services:${SERVICE}:bootBuildImage"
 
   echo "Uploading docker images for ${SERVICE}-service."
   # Login to dockerhubwith credentials