+++ services/client/src/main/java/org/cryptimeleon/incentive/client/IncentiveClientException.java
@@ -1,0 +1,7 @@
+package org.cryptimeleon.incentive.client;
+
+import lombok.AllArgsConstructor;
+
+@AllArgsConstructor
+public class IncentiveClientException extends RuntimeException {
+}