+++ services/client/src/main/java/org/cryptimeleon/incentive/client/dto/PostRedeemBasketDto.java
@@ -1,0 +1,16 @@
+package org.cryptimeleon.incentive.client.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.UUID;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class PostRedeemBasketDto {
+    UUID basketId;
+    String redeemRequest;
+    long value;
+}