+++ services/credit/src/main/java/org/cryptimeleon/incentive/services/credit/interfaces/CreditInterface.java
@@ -1,0 +1,8 @@
+package org.cryptimeleon.incentive.services.credit.interfaces;
+
+/*
+ * This will be replaced by the IncentiveSystem in the future since we changed our project structure.
+ */
+public interface CreditInterface {
+    String computeSerializedResponse(String serializedEarnRequest, long earnAmount);
+}