+++ services/credit/src/test/java/org/cryptimeleon/incentive/services/credit/CreditApplicationTests.java
@@ -1,0 +1,13 @@
+package org.cryptimeleon.incentive.services.credit;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CreditApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}