+++ services/issue/src/main/java/org/cryptimeleon/incentive/services/issue/IncentiveException.java
@@ -1,0 +1,4 @@
+package org.cryptimeleon.incentive.services.issue;
+
+public class IncentiveException extends Exception {
+}